+++ hackernews/lib/screens/splash_screen.dart
@@ -1,0 +1,39 @@
+import 'package:after_layout/after_layout.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:hackernews/screens/home_screen.dart';
+
+class SplashScreen extends StatefulWidget {
+  @override
+  _SplashScreenState createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> with AfterLayoutMixin {
+  @override
+  void afterFirstLayout(BuildContext context) {
+    Future.delayed(
+        Duration(
+          seconds: 3,
+        ), () {
+      Get.to(
+        HomeScreen(),
+      );
+    });
+    print('afterLayout');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text(
+          'Hacker News',
+          style: TextStyle(
+            fontSize: 40,
+            color: Colors.black,
+          ),
+        ),
+      ),
+    );
+  }
+}